Add a tool for Lean frontend extraction (#684)
* Extract ASTs for each subtable MLE and output a representation compatible with the ZkLean library.

We do this by implementing `JoltField` using a type, `MleAst`, that constructs an AST of the computation computed. We then run the `evaluate_mle` function for each implementor of `LassoSubtable`. We print the results as a Lean4 file.

We test that the ASTs are generated correctly by creating an evaluate function for the AST, then writing property tests that compare the result of extracting and evaluating the AST against running `evaluate_mle` on an existing `JoltField` directly.

* Update subtable output to accommodate new ZkLean API

* Resolve "Extract ASTs for combining tables"

* Resolve "Extract R1CS constraints"

* Confirmed that we can treat auxiliaries the same as inputs for R1CS

* Resolve "Extract virtual instructions"

* Fix lookup_step with latest zkLean

The changes are:
- composed tables needed _32 at the end
- comma interspersed in the array of pair instruction_flags,composed_tables
- add a constrainEq between result of mux_lookup and witness LookupOutput

* Generate Witnessable instance for JoltR1CSInputs

* Resolve "Split extraction up into multiple modules"

* Add derive(Debug) to LassoSubtable impls

* Bump ark package versions

* Migrates most of the extraction to using JoltInstruction and LassoSubtable

This commit gets us most of the way to extracting via JoltInstruction and
LassoSubtable. However, pretty-printing valid Lean4 identifiers for the
names is still todo.

* Remove binius dependency from zklean-extractor

* Add static string conversions for JoltInstruction and LassoSubtable

* Make extracted subtable and instruction names into valid Lean identifiers

* Eliminate proc_macro stuff from zklean-extractor and bump Cargo.lock

* Extract constants from Jolt codebase

Jolt uses certain constants for a given instruction set / decomposition
strategy / memory layout. Currently, only one set of constants is supported,
but there may be more sets in the future (e.g., for 64-bit risc-v). This adds
the ability to generalize over those constants and pulls them from the Jolt
codebase, rather than hard-coding them in the extractor.

* Add the jolt step function

* Refactor and add docs

* One more clarifying comment

* Rename JoltField to ZKField for zkLean

* Resolve "Integration tests"

* Fix clippy errors

* Fix some errors caught by graphite

* Run cargo fmt

* Update Cargo.lock

* Fix some typos and linter errors

* Update zklean-extractor README

* Remove unnecessary phantoms

* Remove ark field initialization from tests

The lazy_static library is now being used to intialize the small-value lookup
tables for ark_bn254::Fr, so we don't need to call the initialization function
manually in our tests.

---------

Co-authored-by: Benoit Razet <benoit.razet@galois.com>
Co-authored-by: James Parker <james@galois.com>