Extract operand interleaving for instruction MLEs
The determines how the two input operands of an instruction are combined into a
single vector. Due to the requirement of Twist & Shout that each instruction
MLE be decomposable into prefix and suffix MLEs, some instructions require
their operands to be interleaved, and others concatenated. Jolt determines the
operand interleaving using the circuit flags, so we do the same, here.