Extract circuit flags set for each instruction
Each instruction sets some of the `R1CSInputs::OpFlags` Boolean R1CS input
variables. This extracts those flags. We currently print them in a comment,
pending a decision on how to incorporate them into ZkLean.

Note that *all* instructions conditionally set the `DoNotUpdateUnexpandedPC`
and `InlineSequenceInstruction` flags when they are part of an inline sequence.
We do not extract these, since they depend on the trace, not the instruction,
itself.