Incorporate `Var` and `Scalar` nodes into parent node in `MleAst`
This reduces the number of nodes required in ASTs by putting scalars and
variables in the branches of their parent nodes, rather than making separate
nodes for them. Since manyvariables and constants such as 0 and 1 tend to be
reused many times. The cost of this is that the size of each node is increased,
due to needing to contain space for two scalars, as well as additional overhead
for enum tags. We work on compressing node sizes next.