Reduce size of atomic MLE AST elements
This reduces the size of an AST node by reducing the size of an atomic element
(a scalar or a variable). In particular, we
* store an array of all unique constants in an AST, so that a scalar can be
  represented as an index into this array; and
* store the register name in the AST, rather than in each variable node.
Doing so brings the size of an atom down to 4 bytes and the size of a node down
to 10 bytes (due to the overhead of the enum tags). This is sufficient to
enable all 32-bit lookup tables (which we also do).