Use arena allocation for MLE AST nodes in zklean extractor
Our approach to extracting lookup-table MLEs is to have an AST data
structure that implements `JoltField` and keeps track of the
polynomial computation. However, naively, the `Sized` and `Copy`
constraints on `JoltField` means that we are unable to use variably
sized collections in this data structure, and many ASTs get
allocated on the stack over the course of computations. This runs
into scalability issues, since stack space is limited. This commit
avoids this problem by using a static reference to a locked
allocation arena for AST nodes stored on the heap.