Update R1CS-constraint data structure in extractor
The data structure for R1CS constraints has changed upstream, in particular, allowing R1CS constraints to be expressed in a constant array, rather than a builder pattern. This commit brings the extractor up to date with that data structure.

One thing to note is that the terms and constants of a linear combination within an R1CS constraint does not seem to be directly accessible. For that reason, we make the `LC::decompose` function public in this commit.