Add evaluation tests in Lean for extracted lookup-table MLEs
We test the lookup tables in Lean as follows:
* Generate 2 * WORD_SIZE random field elements to use as input
* For each lookup-table MLE, compute output := evaluate_mle(input)
  in Rust
* Generate a Tests module in Lean with a #guard statement for each
  MLE that checks mle(input) = output
This commit implements the above tests. The field is hard-coded to
be the scalar field of BN254 for now, implemented using ZMod p in
Lean. This requires showing that p is prime, which we sorry out for
now. Proving this or using BN254 from arklib is todo.