Avoid hard-coding list of non-inline-sequence instructions in extractor
When extracting instructions with lookup tables / circuit flags, we
want to avoid inline sequences, because the functions that return
these things panic on inline-sequence instructions. Previously, we
hard-coded the list of supported instructions in the extractor,
which is bad because it requires manual curation. This commit
checks the `inline_sequence` instruction and filters out
instructionst that return non-empty vectors, instead.