Copy BN254.ScalarField from ArkLib for use in extracted Lean tests
This commit allows us to use the BN254 scalar field as if we were
importing it from ArkLib. For now, there is a Lean4 version
mismatch preventing us from actually depending on ArkLib. Once
that's fixed, we can remove our copy and substitute in the ArkLib
dependency without otherwise changing the Lean code.