Implement a streaming version of RLCPolynomial
This commit implements streaming for the `linear_combination` and
`vector_matrix_product` methods of `RLCPolynomial`. The dense and one-hot
implementations are separate for now. Both implementations lazily iterate
over R entries of the polynomial at a time, where R is the size of a row.
Parallelizing the iteration is still todo.