Add a "batching polynomial" field to SumcheckFrontend
This captures polynomials such as Eq, EqPlusOne, Lt, Identity, etc., which
are evaluated by the verifier on the challenge. This allows us to
represent more of the sumchecks. The `batching_poly` field replaces the
prior `is_offset` field, since that information is determined by whether
the batching polynomial is Eq or EqPlusOne.