Generalize the representation of memory inside of the emulator MMU
This adds a `MemoryData` trait for use inside the MMU's `MemoryBackend`.
This allows us to use different data structures to represent RAM within
the emulator. Specifically, we want a data structure that records the
initial state of each accessed entry in memory as a "checkpoint", as well
as one that allows replaying from such a checkpoint.