Implement saving and replaying memory accesses from a hashmap
This adds the `CheckpointingMemory` and `ReplayableMemory` structs, which
can be used as the underlying memory data structure for the emulator. The
former allows saving the initial values of memory accessed during a given
chunk of execution as a hashmap, and the latter uses a hashmap to replay
those memory accesses.