Generalize the memory type in Cpu and Emulator
For now, we rename Emulator to GeneralizedEmulator<D> and Cpu to
GeneralizedCpu<D> and make Emulator and Cpu type aliases, where D is
instantiated as Vec<u64>, as before. This avoids making many upstream
changes to jolt-core, for the time being.

This commit touches many files, due to needing to add generics to all of
the many places Cpu is used, especially the instruction types.