Make read/write API for memory private and add public read-only API
Now that we need to record accesses to memory in the emulator, the read
interface needs to take the memory mutably. This makes it difficult to use
in jolt-core, where the memory needs to be read within a ParIter. In
addition, allowing external modification of the emulator memory is
probably not desirable. Thus, we make the read/write interface private to
the crate and add a collection of simple getters.